Enable dBm/nm and friends!
Having real units output from both the pulse.Pulse and the NLSE.PulseData classes seems very useful. The native units are Watts for both AW and AT, which are mega confusing. So, this PR enables the output of the following units for both the 1D data from the pulse class as well as for the NLSE 2D data. There is some redundant code, so maybe there should be a universal utility function that can deal with both 1D and 2D data.

The options for units are:
- mW/bin
-mW/THz
- dBm/THz
- mW/nm
- dBm/nm

There's a large number of possible units, but it should be straightforward for people to convert to other units (like W/nm) from these values. But, are there other obvious ones that we should include?

I added some unit tests to ensure that integration over the AW and AT arrays with ALL of the unit combinations yields the same pulse energies. Integration over the "per nm" units only agrees to within <1e-6 to integration over the "per bin" or "per THz" units. I'm not precisely sure why this occurs, but it's probably perfectly acceptable for most situations. The purists can stay in frequency land where life is perfect.